Add KrimsonGapTracker
When creating projections in Krimson, its necessary to be able to ensure that a processor has processed all available events in an existing stream. This PR adds an interface IHasCaughtUp that can be used from anywhere DI is available to determine whether a processor with a given client id has caught up.